my-awesome-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiomysql==0.2.0
redis==5.0.1
pillow==10.2.0
loguru==0.7.2
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14